+++ packages/assets-controllers/src/DeFiPositionsController/DeFiPositionsController.ts
@@ -5,12 +5,12 @@ import type {
 import type {
   ControllerGetStateAction,
   ControllerStateChangeEvent,
-  RestrictedMessenger,
   StateMetadata,
-} from '@metamask/base-controller';
+} from '@metamask/base-controller/next';
 import { isEvmAccountType } from '@metamask/keyring-api';
 import type { KeyringControllerLockEvent } from '@metamask/keyring-controller';
 import type { InternalAccount } from '@metamask/keyring-internal-api';
+import type { Messenger } from '@metamask/messenger';
 import { StaticIntervalPollingController } from '@metamask/polling-controller';
 import type { TransactionControllerTransactionConfirmedEvent } from '@metamask/transaction-controller';
 import type { Hex } from '@metamask/utils';
@@ -68,13 +68,13 @@ const controllerMetadata: StateMetadata<DeFiPositionsControllerState> = {
   allDeFiPositions: {
     includeInStateLogs: false,
     persist: false,
-    anonymous: false,
+    includeInDebugSnapshot: false,
     usedInUi: true,
   },
   allDeFiPositionsCount: {
     includeInStateLogs: false,
     persist: false,
-    anonymous: false,
+    includeInDebugSnapshot: false,
     usedInUi: false,
   },
 };
@@ -121,12 +121,10 @@ export type AllowedEvents =
 /**
  * The messenger of the {@link DeFiPositionsController}.
  */
-export type DeFiPositionsControllerMessenger = RestrictedMessenger<
+export type DeFiPositionsControllerMessenger = Messenger<
   typeof controllerName,
   DeFiPositionsControllerActions | AllowedActions,
-  DeFiPositionsControllerEvents | AllowedEvents,
-  AllowedActions['type'],
-  AllowedEvents['type']
+  DeFiPositionsControllerEvents | AllowedEvents
 >;
 
 /**
@@ -174,11 +172,11 @@ export class DeFiPositionsController extends StaticIntervalPollingController()<
     this.#fetchPositions = buildPositionFetcher();
     this.#isEnabled = isEnabled;
 
-    this.messagingSystem.subscribe('KeyringController:lock', () => {
+    this.messenger.subscribe('KeyringController:lock', () => {
       this.stopAllPolling();
     });
 
-    this.messagingSystem.subscribe(
+    this.messenger.subscribe(
       'TransactionController:transactionConfirmed',
       async (transactionMeta) => {
         const selectedAddress = this.#getSelectedEvmAdress();
@@ -195,7 +193,7 @@ export class DeFiPositionsController extends StaticIntervalPollingController()<
       },
     );
 
-    this.messagingSystem.subscribe(
+    this.messenger.subscribe(
       'AccountTreeController:selectedAccountGroupChange',
       async () => {
         const selectedAddress = this.#getSelectedEvmAdress();
@@ -287,7 +285,7 @@ export class DeFiPositionsController extends StaticIntervalPollingController()<
   }
 
   #getSelectedEvmAdress(): string | undefined {
-    return this.messagingSystem
+    return this.messenger
       .call('AccountTreeController:getAccountsFromSelectedAccountGroup')
       .find((account: InternalAccount) => isEvmAccountType(account.type))
       ?.address;